+++ lib/atoms.ts
@@ -1,0 +1,4 @@
+import { atom } from "jotai";
+import { getDefaultSettings } from "./types";
+
+export const settingsAtom = atom(getDefaultSettings())